+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Removed
+- [**BREAKING**] `YoleckLoadingCommand::FromData` is removed. If it is
+  necessary to load a level from memory, add it to
+  `ResMut<Assets<YoleckRawLevel>>` first and pass the handle to
+  `YoleckLoadingCommand::FromAsset`.
 
 ## 0.16.0 - 2023-09-06
 ### Changed